# Conflicts:
#	smart-contract/js/tests/end_to_end.test.ts
#	smart-contract/program/src/processor/activate_stake_pool.rs
#	smart-contract/program/src/processor/claim_bond_rewards.rs
#	smart-contract/program/src/processor/crank.rs
#	smart-contract/program/src/processor/create_bond.rs
#	smart-contract/program/src/processor/execute_unstake.rs
#	smart-contract/program/src/processor/stake.rs
#	smart-contract/program/src/processor/unstake.rs
#	smart-contract/program/src/state.rs
#	smart-contract/program/src/utils.rs
#	smart-contract/program/tests/common/test_runner.rs
#	smart-contract/program/tests/functional.rs
#	smart-contract/program/tests/repeated_claim.rs
#	smart-contract/program/tests/rewards_bonds.rs

+++ smart-contract/program/src/processor/claim_bond_rewards.rs
@@ -145,7 +145,7 @@ pub fn process_claim_bond_rewards(
 
     let reward = calc_reward_fp32(
         central_state.last_snapshot_offset,
-        bond.last_claimed_offset as i64,
+        bond.last_claimed_offset,
         &stake_pool,
         true,
         false,
